Migrate yaml.
