+++ crates/cli/Cargo.toml
@@ -50,7 +50,6 @@ serde_yaml = { workspace = true }
 strum = { version = "0.24.1", features = ["derive"] }
 tera = { version = "1.17.1", features = ["preserve_order"] }
 tokio = { workspace = true }
-yaml-rust = { version = "0.5.1", package = "yaml-rust-davvid" }
 
 [dev-dependencies]
 moon_archive = { path = "../core/archive" }
