+++ crates/core/utils/Cargo.toml
@@ -32,4 +32,5 @@ serde_yaml = { workspace = true }
 shell-words = "1.1.0"
 tokio = { workspace = true }
 wax = "0.5.0"
+yaml-rust = { version = "0.5.1", package = "yaml-rust-davvid" }
