+++ crates/node/lang/src/pnpm/mod.rs
@@ -6,7 +6,7 @@ use cached::proc_macro::cached;
 use dependency_path::PnpmDependencyPath;
 use moon_error::MoonError;
 use moon_lang::{config_cache, LockfileDependencyVersions};
-use moon_utils::fs::sync::read_yaml;
+use moon_utils::yaml::read as read_yaml;
 use rustc_hash::FxHashMap;
 use serde::{Deserialize, Serialize};
 use serde_yaml::Value;
