+++ crates/node/lang/src/pnpm/workspace.rs
@@ -2,7 +2,7 @@ use crate::PNPM;
 use cached::proc_macro::cached;
 use moon_error::MoonError;
 use moon_lang::config_cache;
-use moon_utils::fs::sync::read_yaml;
+use moon_utils::yaml::read as read_yaml;
 use rustc_hash::FxHashMap;
 use serde::{Deserialize, Serialize};
 use serde_yaml::Value;
